blocksync: moved proto aliases from api package
